init:refactor: Change name of front end packages to adhere to AD001 until AD002
