+++ src/main/java/be/archilios/wishingwell/frontend/MainLayout.java
@@ -1,8 +1,8 @@
-package be.archilios.wishingwell.views;
+package be.archilios.wishingwell.frontend;
 
 
-import be.archilios.wishingwell.views.about.AboutView;
-import be.archilios.wishingwell.views.helloworld.HelloWorldView;
+import be.archilios.wishingwell.frontend.about.AboutView;
+import be.archilios.wishingwell.frontend.helloworld.HelloWorldView;
 import com.vaadin.flow.component.Component;
 import com.vaadin.flow.component.applayout.AppLayout;
 import com.vaadin.flow.component.html.Div;
